Split feature init to 3 phases to prepare for VMs (#44)
First, the instance is started so that the initialization phase can run
commands in it. Once the initialization phase (e.g. useradd + idmap) is
done, the instance is restarted. After that, comes the configuration
phase where most features configure themselves. Finally, the instance is
stopped and the new instance configurations are applied.

- (start)
- INIT
- (restart)
- CONFIGURE
- (stop)
- UPDATE