Settings: rename IDs for future-proofing (#216)
Earlier, we were forced to use hyphens because of GSettings. Now that we
have a JSON backend, we can switch to VSCode-like IDs that will allow
implementing support for an alternative structural settings syntax:

    "terminal.fontFamily": "Ubuntu Mono",
    "terminal.fontSize": 16.0

vs.

    "terminal": {
      "fontFamily": "Ubuntu Mono",
      "fontsize": 16.0
    }

Notice that `jsettings.dart` does not (yet) support the latter syntax. This is just
a hypothetical preparation step to have such IDs that make it possible
to support this later on.

NOTE: There's no any sort of migration so this breaks any existing user
settings (just like we did with the GSettings to JSON switch yesterday).