simplestreams: drop freezed in favor of equatable
The simplestreams package has proven itself and could be ready for
promotion to a standalone package in its own repo.

This change prepares for that by dropping the heavy freezed dependency
from the public API by replacing it with a much lighter alternative,
equatable. This results in a bit more repeat but significantly less
generated code.

The only thing we lose with this trade is copyWith(), which is unlikely
to be needed for such read-only data as the simplestreams information.