Remove malfunctioning command palette close shortcut
First of all, the shortcut did not work as expected. Escape works for
closing the command palette regardless of what we try to define in the
page context because the command palette is opened as an overlay into a
different context.

Secondly, defining Escape shortcut on this level is problematic because
it eats Escape key presses sent to the terminal pane.

Fixes: #311