Agressively upgrades to Flutter 3.24
Which requires Dart SDK 3.5 and intl 0.19.0.
ubuntu_localizations 0.5.1 is incompatible with Flutter 3.27.
The fix seems partially implemented in the upstream repo, not yet
released.
More fixes to come.
Some other packages are upgraded because of broken dependencies.
With those upgrades we can now run `melos boostrap` and `melos analyze`,
although the latest raises a bunch of warnings.
Next tasks are:
- [ ] Fix issues raised by `melos analyze`
- [ ] Fix `melos test` errors
- [ ] Fix snap build in CI